% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{add_package_access}
\alias{add_package_access}
\title{Add standard access rules for an object}
\usage{
add_package_access(
  pid,
  subjects,
  permissions = c("read", "write", "changePermission"),
  repo = "urn:node:KNB"
)
}
\arguments{
\item{pid}{(character) The PID identifier of the object to check permissions for.}

\item{subjects}{(character) List of subjects to be checked for the permissions}

\item{permissions}{(character) List of permissions (read, write, changePermission) to be checked}

\item{repo}{(character or MNode) The identifier of the repository as a string value, in which case it will be instantiated,
or as an instance of \code{dataone::MNode}}
}
\value{
(logical) Whether access rules needed to be added.
}
\description{
Add the access policy rules for the given subjects for the given objects on a repository.
}
\examples{
\dontrun{
pid <- "resource_map_urn:uuid:e788f9f8-8990-4440-882f-9f65deaa69f2"
ginny <- "http://orcid.org/0000-0002-7538-7081"
sme_group <- "CN=Sustainable Marine Ecosystems in Alaska Lab,DC=dataone,DC=org"
add_package_access(pid, subjects = c(ginny, sme),
    permissions = c("read", "write", "changePermission"), repo="urn:node:KNB")
}
}
