% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{check_package_access}
\alias{check_package_access}
\title{Check and report access rules for a data package}
\usage{
check_package_access(
  pid,
  subjects,
  permissions = c("read", "write", "changePermission"),
  repo = "urn:node:KNB"
)
}
\arguments{
\item{pid}{(character) The PID identifier of the package to check permissions for.}

\item{subjects}{(character) List of subjects to be checked for the permissions}

\item{permissions}{(character) List of permissions (read, write, changePermission) to be checked}

\item{repo}{(character or MNode) The identifier of the repository as a string value, in which case it will be instantiated,
or as an instance of \code{dataone::MNode}}
}
\value{
(data.frame) Listing the status of the access rules for each member of the package.
}
\description{
Check the access policy for the given subjects for the given package on a repository.
}
\examples{
\dontrun{
pid <- "resource_map_urn:uuid:e788f9f8-8990-4440-882f-9f65deaa69f2"
ginny <- "http://orcid.org/0000-0002-7538-7081"
sme_group <- "CN=Sustainable Marine Ecosystems in Alaska Lab,DC=dataone,DC=org"
check_package_access(pid, subjects = c(ginny, sme_group),
    permissions = c("read", "write", "changePermission"))

}
}
