% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{set_access}
\alias{set_access}
\title{Set the access policy for an object}
\usage{
set_access(
  pids,
  subjects,
  permissions = c("read", "write", "changePermission"),
  repo = "urn:node:KNB"
)
}
\arguments{
\item{pids}{(character) The PIDs of the objects to set permissions for.}

\item{subjects}{(character) The identifiers of the subjects to set permissions for, typically an ORCID or DN.}

\item{permissions}{(character) Optional. The permissions to set. Defaults to
read, write, and changePermission.}

\item{repo}{(character or MNode) The identifier of the repository as a string value, in which case it will be instantiated,
or as an instance of \code{dataone::MNode}}
}
\value{
(logical) Whether an update was needed.
}
\description{
Set the access policy for the given subjects for the given objects on the given Member Node.
For each type of permission, this function checks if the permission is already set
and only updates the System Metadata when a change is needed.
}
\examples{
\dontrun{
pids <- c("urn:uuid:3e5307c4-0bf3-4fd3-939c-112d4d11e8a1",
          "urn:uuid:23c7cae4-0fc8-4241-96bb-aa8ed94d71fe")
set_access(pids, subjects = ginny,
           permissions = c("read", "write", "changePermission"))
}
}
